Add reader/writer for download data from and uploading data to server
